+++ CalculatorProject/src/operator/DivideOperator.java
@@ -5,7 +5,7 @@ import exception.DivisionByZeroException;
 public class DivideOperator implements UserOperator {
 
     @Override
-    public int calculate(int firstValue, int secondValue) throws DivisionByZeroException {
+    public double calculate(double firstValue, double secondValue) throws DivisionByZeroException {
         if (secondValue == 0) {
             throw new DivisionByZeroException("0으로 나눌 수 없습니다.");
         }
