+++ CalculatorProject/src/operator/UserOperator.java
@@ -4,5 +4,5 @@ import exception.DivisionByZeroException;
 
 public interface UserOperator {
 
-    int calculate(int firstValue, int secondValue) throws DivisionByZeroException;
+    double calculate(double firstValue, double secondValue) throws DivisionByZeroException;
 }
